requirements.txt
pdf2image
openai
Pillow
python-docx
